+++ contrib/vtkSurfaceReconstructionFilter.cxx
@@ -398,8 +398,8 @@ void vtkSurfaceReconstructionFilter::Execute()
   if(this->SampleSpacing<=0.0)
 	{
 	  // sample spacing guessed as cube root of (volume divided by number of points)
-	  this->SampleSpacing = pow((bounds[1]-bounds[0])*(bounds[3]-bounds[2])*(bounds[5]-bounds[4]) / (float)COUNT,
-		  1.0/3.0);
+	  this->SampleSpacing = pow((double)(bounds[1]-bounds[0])*(bounds[3]-bounds[2])*(bounds[5]-bounds[4]) / (float)COUNT,
+		  (double)(1.0/3.0));
  
 	  vtkDebugMacro(<<"Estimated sample spacing as: " << this->SampleSpacing );
 	}
